requirements.txt
gunicorn==20.1.0
pyodbc
flask_session
sqlalchemy
python-dotenv
flask
